+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [#2883](https://github.com/influxdata/telegraf/pull/2883): Add support for an optional SSL/TLS configuration to nginx input plugin
 - [#2882](https://github.com/influxdata/telegraf/pull/2882): Add timezone support for logparser timestamps.
 - [#2814](https://github.com/influxdata/telegraf/pull/2814): Add result_type field for http_response input.
+- [#2734](https://github.com/influxdata/telegraf/pull/2734): Add include/exclude filters for docker containers.
 
 ### Bugfixes
 
@@ -33,6 +34,7 @@
 ### Bugfixes
 
 - [#2862](https://github.com/influxdata/telegraf/issue/2862): Fix InfluxDB UDP metric splitting.
+- [#2888](https://github.com/influxdata/telegraf/issue/2888): Fix mongodb/leofs urls without scheme.
 
 ## v1.3.1 [2017-05-31]